Restore stock keyframe editor: drop my layout/visual changes
Reverts this session's keyframe editor layout/visuals to the BBS Fresh baseline
(06fad365): keyframe size, recolour unused track lines, hide track lines, the
left/right label toggle, settable properties-panel height, and the properties-
below-goes-live change. The visual-only editor files are restored to baseline;
UIKeyframeClip and BBSSettings edited surgically. Kept: live pose recording, the
timeline-jump fixes, secondary anchor, and the Films button.

+++ src/client/java/mchorse/bbs_mod/ui/framework/elements/input/keyframes/shapes/CircleKeyframeShapeRenderer.java
@@ -23,7 +23,7 @@ public class CircleKeyframeShapeRenderer implements IKeyframeShapeRenderer
     }
 
     @Override
-    public void renderKeyframe(UIContext uiContext, BufferBuilder builder, Matrix4f matrix, int x, int y, float offset, int c)
+    public void renderKeyframe(UIContext uiContext, BufferBuilder builder, Matrix4f matrix, int x, int y, int offset, int c)
     {
         final int NUM_SEGMENTS = 32;
 
@@ -57,7 +57,7 @@ public class CircleKeyframeShapeRenderer implements IKeyframeShapeRenderer
     }
 
     @Override
-    public void renderKeyframeBackground(UIContext uiContext, BufferBuilder builder, Matrix4f matrix, int x, int y, float offset, int c)
+    public void renderKeyframeBackground(UIContext uiContext, BufferBuilder builder, Matrix4f matrix, int x, int y, int offset, int c)
     {
         float centerSize = offset * 0.2f;
         float half = centerSize * 2;
